# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server for Northwind database"
  version: "1.0.0"
  title: "Northwind database"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "danylomaster@ukr.net"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:5000"
basePath: "/api/v1"
...
schemes:
- "http"
- "https"
paths:
  /customers/:
    get:
      tags:
      - "customers"
      summary: "List of all customers"
      description: ""
      operationId: "getCustomers"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Customer"
      x-swagger-router-controller: "Customers"
    post:
      tags:
      - "customers"
      summary: "Create a customer in the database with form data"
      description: ""
      operationId: "createCustomer"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "company"
        in: "formData"
        description: "Updated company of the customer"
        required: false
        type: "string"
      - name: "last_name"
        in: "formData"
        description: "Updated last_name of the customer"
        required: false
        type: "string"
      - name: "first_name"
        in: "formData"
        description: "Updated first_name of the customer"
        required: false
        type: "string"
      - name: "email_address"
        in: "formData"
        description: "Updated email_address of the customer"
        required: false
        type: "string"
      - name: "job_title"
        in: "formData"
        description: "Updated job_title of the customer"
        required: false
        type: "string"
      - name: "business_phone"
        in: "formData"
        description: "Updated business_phone of the customer"
        required: false
        type: "string"
      - name: "home_phone"
        in: "formData"
        description: "Updated home_phone of the customer"
        required: false
        type: "string"
      - name: "mobile_phone"
        in: "formData"
        description: "Updated mobile_phone of the customer"
        required: false
        type: "string"
      - name: "fax_number"
        in: "formData"
        description: "Updated fax_number of the customer"
        required: false
        type: "string"
      - name: "address"
        in: "formData"
        description: "Updated address of the customer"
        required: false
        type: "string"
      - name: "city"
        in: "formData"
        description: "Updated city of the customer"
        required: false
        type: "string"
      - name: "state_province"
        in: "formData"
        description: "Updated state_province of the customer"
        required: false
        type: "string"
      - name: "zip_postal_code"
        in: "formData"
        description: "Updated zip_postal_code of the customer"
        required: false
        type: "string"
      - name: "country_region"
        in: "formData"
        description: "Updated country_region of the customer"
        required: false
        type: "string"
      - name: "web_page"
        in: "formData"
        description: "Updated web_page of the customer"
        required: false
        type: "string"
      - name: "notes"
        in: "formData"
        description: "Updated notes of the customer"
        required: false
        type: "string"
      - name: "attachments"
        in: "formData"
        description: "Updated attachments of the customer"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Customers"
  /customers/{customerId}:
    get:
      tags:
      - "customers"
      summary: "Find customer by ID"
      description: "Returns a single Customer"
      operationId: "getCustomerById"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "ID of Customer to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Customer"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Customer not found"
      x-swagger-router-controller: "Customers"
    post:
      tags:
      - "customers"
      summary: "Updates a customer in the store with form data"
      description: ""
      operationId: "updateCustomer"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "ID of customer that needs to be updated"
        required: true
        type: "integer"
      - name: "company"
        in: "formData"
        description: "Updated company of the customer"
        required: false
        type: "string"
      - name: "last_name"
        in: "formData"
        description: "Updated last_name of the customer"
        required: false
        type: "string"
      - name: "first_name"
        in: "formData"
        description: "Updated first_name of the customer"
        required: false
        type: "string"
      - name: "email_address"
        in: "formData"
        description: "Updated email_address of the customer"
        required: false
        type: "string"
      - name: "job_title"
        in: "formData"
        description: "Updated job_title of the customer"
        required: false
        type: "string"
      - name: "business_phone"
        in: "formData"
        description: "Updated business_phone of the customer"
        required: false
        type: "string"
      - name: "home_phone"
        in: "formData"
        description: "Updated home_phone of the customer"
        required: false
        type: "string"
      - name: "mobile_phone"
        in: "formData"
        description: "Updated mobile_phone of the customer"
        required: false
        type: "string"
      - name: "fax_number"
        in: "formData"
        description: "Updated fax_number of the customer"
        required: false
        type: "string"
      - name: "address"
        in: "formData"
        description: "Updated address of the customer"
        required: false
        type: "string"
      - name: "city"
        in: "formData"
        description: "Updated city of the customer"
        required: false
        type: "string"
      - name: "state_province"
        in: "formData"
        description: "Updated state_province of the customer"
        required: false
        type: "string"
      - name: "zip_postal_code"
        in: "formData"
        description: "Updated zip_postal_code of the customer"
        required: false
        type: "string"
      - name: "country_region"
        in: "formData"
        description: "Updated country_region of the customer"
        required: false
        type: "string"
      - name: "web_page"
        in: "formData"
        description: "Updated web_page of the customer"
        required: false
        type: "string"
      - name: "notes"
        in: "formData"
        description: "Updated notes of the customer"
        required: false
        type: "string"
      - name: "attachments"
        in: "formData"
        description: "Updated attachments of the customer"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Customers"
    delete:
      tags:
      - "customers"
      summary: "Deletes a customer"
      description: ""
      operationId: "deleteCustomer"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "Customer id to delete"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Customer not found"
      x-swagger-router-controller: "Customers"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Customer:
    type: "object"
    properties:
      id:
        type: "integer"
      company:
        type: "string"
      last_name:
        type: "string"
      first_name:
        type: "string"
      email_address:
        type: "string"
      job_title:
        type: "string"
      business_phone:
        type: "string"
      home_phone:
        type: "string"
      mobile_phone:
        type: "string"
      fax_number:
        type: "string"
      address:
        type: "string"
      city:
        type: "string"
      state_province:
        type: "string"
      zip_postal_code:
        type: "string"
      country_region:
        type: "string"
      web_page:
        type: "string"
      notes:
        type: "string"
      attachments:
        type: "string"
    example:
      home_phone: "home_phone"
      address: "address"
      notes: "notes"
      attachments: "attachments"
      city: "city"
      zip_postal_code: "zip_postal_code"
      last_name: "last_name"
      state_province: "state_province"
      fax_number: "fax_number"
      country_region: "country_region"
      web_page: "web_page"
      business_phone: "business_phone"
      email_address: "email_address"
      mobile_phone: "mobile_phone"
      company: "company"
      id: 0
      first_name: "first_name"
      job_title: "job_title"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
      type:
        type: "string"
      message:
        type: "string"
    example:
      code: 0
      type: "type"
      message: "message"
...